pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
version = "0.0.0"
[tool.poetry.dependencies]
"discord.py" = "==0.16.12"
flask = "*"
mcstatus = "*"
python = "^3.7"
